Get rid of crlf, favoring CR LF
The abbreviations for carriage return and newline are CR LF (in the
list which includes NUL for the 0 character).  These are common enough
to likely warrant being defined by default, and should probably be
capitalized when they are used.

However, the string containing both characters was defined as `crlf`.
This takes already lousy abbreviations and endorses the use of them
in combination.  The logic of SPACED does not put delimiters when
characters are used, so:

    spaced ["some" "text" CR LF "more" "text" CR LF]

Would produce "some text" on a line, and then "more text" on a line,
with CR LF between them and no additional spacing.  This is better
than having a string for CRLF, which would be treated with spacing
logic, giving a space after both "text" and before "more".

Also improves/corrects some comments mixing up UCS-2 and UTF-16, and
updates SET and GET in r3-legacy.