Shim stdbool to use false and true, fix bugs found by it
While C prior to C99 did not have a boolean datatype, there is little
to be gained by using `TRUE` and `FALSE` in the source instead of
`true` and `false`.  One simply has to determine if the language has
a true and false definition in it already, and if not, don't redefine
as 0 and 1.

This offers immediate benefit in terms of catching cases where integer
assignments are truncated as booleans.  There were a couple of bugs
of this nature found by the change.  It also basically eliminates the
need for DEBUG_STRICT_BOOL, since C++ compilers warn on integer
truncation to bool.

A global search-and-replace for TRUE and FALSE can be done at some
point to use TRUE and FALSE instead.  This commit does not do that yet.