Native-optimized OR/AND/XOR, ENFIX in native C-comment specs
While they were being designed, it was convenient to have the infix
boolean operations be written in Rebol.  However, this made them
significantly less efficient than ANY or ALL in practice.

Since the behavior has become fairly settled now, this goes ahead and
implements them as natives in C.  It also adds the ability for natives
to express their wish to be enfixed as part of bootstrap

This makes the bootstrap process which loads natives able to understand
the syntax `foo: enfix native [...]`, so that natives can be written
directly as enfix without needing usermode code to update their
binding.  As a result, a hack that was setting ACTION specifically
to enfix in bootstrap could be removed.