Slight reworking of PARSE-ASN
This makes some tweaks to the PARSE-ASN function in the TLS code,
to get it to use more datatypes.  It uses TAG! for the ASN tags, an
ISSUE! for the current mode it is processing, and an EMAIL! for the
ASN class.

(The idea is that the class would have leading @, such as @universal
or @application.  But bootstrap cannot support this at the moment,
so it uses a trailing @ as a memo to update it when it can be updated.)

It also uses COLLECT to avoid an arbitrary 16-length BLOCK!, as well
as KEEP/LINE to get more readable output from the parse.