Enhance TLS "emit" to preserve buffer positions
The EMIT function in the TLS code would add data to the context's
binary message buffer.  However, it was sometimes necessary to go back
in and patch this data after the emit, to account for the ultimate
length of the message.  The code doing this was hard to follow, so
this commit enhances EMIT such that any SET-WORD! in the body of
the message being formed are set to the position of the output buffer
at that moment in the emit.

Also adds -> as a complement to <-