Make HELP and SOURCE work for all lib definitions, add to tests
HELP and SOURCE have become more complex in Ren-C, due to the
appearance of function compositions.  It attempts to have their "meta"
information guide the help system--to try and explain relationships,
while not duplicating unnecessary information.

(e.g. a function with 20 parameters that is specialized to have one
parameter fixed, should not need all other 19 parameter descriptions to
be copied onto a new function)

Additionally, it exercises the FRAME! concept some, by using the
natural frame layout of the function to store fields...to theoretically
make it easier to reflect the function data without parsing a spec
block (which becomes intractable/impossible when it's a composition).

The code trying to sort out these relationships has come along
organically, and involves a number of core questions--such as when to
use null vs. blank.  It has had a tendency to break.  This commit adds
tests that run SOURCE and HELP on everything in lib...and includes the
necessary fixes to make that happen without error.

(This is not to say that the output, e.g. for source, is useful or
correct at this point.  Just that it doesn't error.)

The tests are included even though they spew a large amount of output.
While the principle of the tests is that they don't have output as
part of their operations, at the moment the pragmatic matter of keeping
the functions working is more important.