Make CSCAPE handle indented substitutions
If a substitution being made into a CSCAPE string spans multiple lines,
this makes sure that all the ensuing lines are at the same indentation
level as where the first one starts.

One application of this is in the BINARY-TO-C conversions, which made
many lines of hex character literals to be linked in as constants.
It also took the compression percentage announcement out of the
make process and put it into comments, as reading it as it went by
was not very actionable information for the build.