Allow AS for more datatypes, BLANK! input to AS/TO
This pre-emptively changes AS to be usable for more datatypes, e.g.
to alias strings and binaries or strings and words...even though the
mechanics aren't actually in place to do that legitimately until
after UTF-8 everywhere.

Instead, it provides a weaker contract.  If you try and AS read-only
data as a for-the-moment incompatible type, it will make a read-only
copy in the new format and position you on that.  Then, if you try and
AS mutable data, it will make a copy in the new format but trash the
original data so trying to read or write it after the aliasing point
will fail.