FREE operation on series data with outstanding references
New capability:

    >> foo: collect [loop 1000 [keep <x>]]

    >> bar: foo

    >> bar/100
    == <x>

    >> free foo

    >> bar/100
    ** Error: Series data has been freed

A small amount of memory (a series node) still hangs around. That’s 8
pointers in size, and won’t go away until all the references are gone.
But the arbitrarily-large data allocation is freed up. If you have a
BINARY! blob that was a megabyte in size, you will only have 32 bytes
remaining once you FREE it (or 64 bytes on 64-bit platforms).

The flag controlling this--SERIES_INFO_INACCESSIBLE--already existed.
It marked when a FRAME! for a function was no longer on the stack, so
the data was gone. But it has been promoted to a general feature of any
series. This means referencing data is a little slower because it has
to check for the flag, though this could be further optimized to
a canon info bits pattern in the future.

Semantically speaking, it's generally the case that there is no
legitimate case for accessing freed data.  However, some code like the
console wants to be able to present an error regarding the free state
without raising an error itself.  This is handled with `FREE?`