Kill SPELLING-OF in favor of AS TEXT!, COPY AS TEXT!
The AS operator now allows one to alias the series data of any string
or word type to any other.  Saying AS TEXT! hence works on ANY-WORD!
types to get their "spelling", and COPY AS TEXT! will get a unique
copy.  While the proxying of words to strings is currently artificial,
with UTF-8 Everywhere it will not be...so this is the preferred
way to get the spelling.