Add ENFIXED? testing to PATH!
While PATH! does not currently do enfix dispatch, this changes the
path lookup to preserve the enfix flag when retrieved from a reference.

This also corrects the case of when HELP is done on an ERROR!, which
due to use of TRAP was being aliased as a variable that was not set.

See: https://github.com/rebol/rebol-issues/issues/2308