Fix BREAK_NOW() macro
The BREAK_NOW() macro wasn't wrapped in something like a do...while(),
and as a result if you said **if (...) BREAK_ON_TICK()** it would
wind up only taking the first line of the macro when the if clause
was true, and run the rest unconditionally.

Includes minor HELP and test fixes.