Patch bug in DEVOID use on END cells, REMOLD=>adaptation
The DEVOID operation is used to migrate void cell pointers to raw NULL
C pointers when they are being exposed to the API.  The variadic
API call used to invoke loops with arguments has some loop forms that
don't pass condition arguments (e.g. LOOP has no condition evaluation
to pass) so it only accepts arity-0 functions...hence the variadic
argument gets an END marker.  END markers were not tolerated by
DEVOID so this caused a crash.

This uses VAL_TYPE_OR_0 instead of IS_VOID to accomodate the end marker
in the devoiding process.

Adds a small test for the function variant of LOOP.  Also changes
the REMOLD to operate as an adaptation, which simplifies it (though
with a name like REMOLD, it's likely to be legacy-only at some point).