PATH!s with inert head elements are themselves inert
This makes it so that any evaluator-inactive type at the head of a path
will curb evaluation of that path.  Hence you can write:

    >> <json>/1.2.30
    == <json>/1.2.30

    >> /dir/subdir ;-- first path element is REFINEMENT!
    == /dir/subdir

    >> "abc"/2
    == "abc"/2

The path can become activated by either putting the head element in
a GROUP!, or referring to it via an evaluative entity

    >> ("abc")/2
    == #"b"

    >> text: "abc"
    >> text/2
    == #"b"

Since ANY-WORD! historically has no path behavior, it was difficult
to demonstrate the inert vs. active difference.  This adds a quick and
dirty ability to pick characters by integer out of words--which with
the ANY-STRING!/ANY-WORD! unification would ultimately align with
whatever other operations might be legal on strings.