Fix nested LOCK of MAP!, fix -> bug on END, module options
This contains some fixes to issues discovered while working on the
ZeroMQ extension.  One issue was that recursive locking was not
properly handling MAP! since it does not allow VAL_SERIES() to extract
its pairlist, nor is it considered an ANY_SERIES().  So it has to be
checked for and handled explicitly.

Another issue is that the hack used to make it so that <- is able to
stop left enfix consumption (marking it as END-tolerant) did not match
the spec, which required it to not return NULL.  So if it ever actually
did encounter an END, an assert would be hit.  This manually raises an
error vs. continuing and returning a NULL--although the hack needs to
be reconsidered at some point.

Also, if a module did not have an Options: section, an error was
being raised from a path access that was yielding null.  This is in
a category of errors that likely won't exist soon with new rules for
NULLs...but it is patched here anyway.