Allow actions to return <opt> any-value! by default
To try and help people writing FUNCTIONs (as opposed to PROCEDUREs)
be sure they were "actually returning a value", nulls were disallowed
by default.  To make it more obvious when an action was being returned
(since callers would have a different experience with them), returning
an ACTION! was disallowed by default as well.

It was always a speculative feature, with the belief that the worst
case scenario is that it would make more functions document their
return types.  If it were deemed too much, then it would have caused
some benefit in the meantime.

All told, it probably caused as much hassle as help.  And since NULL
is emerging to take BLANK!s place as the standard way of saying "no
result", it really doesn't make a lot of sense to prohibit it.

So this just goes back to allowing all types by default.  (The limit
on null *arguments* by default remains, and for now also a limit on
ACTION! arguments so one need not worry about accessing them via
GET-WORD! unless they are explicitly promoted to ANY-VALUE!)