PROC/PROCEDURE set return type as [<opt>] instead of leaving blank
A distinction between BLANK! and unset was previously used to tell the
difference between "returns no result" (e.g. a procedure) or that
the return result was undocumented.  This created a fair bit of
confusion, and it's a lot better if the PROC generator goes ahead and
sets the meta information to say that the return type is `[<opt>]`,
which makes things more consistent.

Not doing null assignments in object creations is also a precursor to
a new rule for nulls, that they can't be used to unset variables any
longer.  The rule is more similar to R3-Alpha's rule for UNSET!, but
with significant twists regarding null's general usage pattern and
non-value status.