Use OF in bootstrap, make source-analysis use null convention
Since the baseline commit is able to to quoted left enfix, this means
plain OF can be used as in `length of`.  This looks better than
`length-of` so the pattern is applied to bootstrap files.

Also does some light updates of the source analysis code for the new
null conventions.  In the process, also adds some comments to it
and changes it so that it uses TAG! instead of WORD! for the failed
check label.

Includes a test for PICK out of a PAIR!, from an issue that had been
resolved on GitHub.