Simplify ANY + ALL + NONE, make ALL [] return NULL
Nulls (Voids) had previously been considered "neither true nor false".
While they couldn't be used directly in conditionals like IF or CASE,
the ANY and ALL functions would ignore them.  This was speculated as
being a possible kind of "tristate" feature, where a function called
by an ANY or ALL could choose to vote yes, vote no, or abstain.

Yet the initial reason the feature existed was to continue the ability
from Rebol2 and R3-Alpha for non-result-returning functions (like
PRINT) to run in mid-ANY or mid-ALL.  With the emergence of invisibles
like ELIDE and DUMP, this was no longer as compelling as having
reliable reactions to NULLs...which as they became more common, should
be errors in all conditionals.

The feature of ignoring nulls was not immediately deleted, just moved
to an /OPT refinement of ANY and ALL.  Using it restored the null
ignoring behavior that had been there previously, as opposed to errors.

However, when NULL was converted to being falsey, there isn't any
difference for an ANY that ignores them vs one that treats them as
falsey.  And the ALL behavior of ignoring nulls didn't have any
remaining usages.  So this goes ahead and gets rid of the /OPT, which
simplifies the code noticeably.

This also resolves an age-old question of what ALL [] should be.  It
had been true in Rebol2 and R3-Alpha, but none in Red (while `ALL [()]`
in Red is unset).  The most sensible answer after the null change is
to move to where `all [] = null`, since doing otherwise would mean
synthesizing some truthy value from mid-air.  There is no good
reason to do that, since no one would write all [] literally in source
any way, so the only reason it would happen would be in generated
code, that could generate its own result in this case, e.g. either
`all [...] else [...]` or `all [... true]`.