Eliminate notion of "unmanaged values", improves performance
Issue #2 in Ren-C was to resolve a problem in R3-Alpha that had a "TBD"
note, regarding dealing with orphan series that weren't ready to be
seen by the garbage collector:

https://github.com/metaeducation/ren-c/issues/2

This eventually became the notion of a "manual" series...one that the
user had never seen, so it could be automatically freed in the case
of an error, or manually freed if the system decided it didn't actually
need it.  But it was a one-way street...once a series was managed, the
only thing that could free it would be the GC, after it had determined
no references existed.

To make matters somewhat easier, putting an unmanaged series into a
cell would ensure it was managed.  So Init_Array() when given a REBARR*
that was not managed would manage it.  Yet it was still mechanically
possible to create an "unmanaged value", and sometimes this had to be
done e.g. for debug probing of a series...for instance because Mold
required a REBVAL* and couldn't just render a plain REBSER*.

This meant that in various parts of the system there were "unmanaged
values"...series that lived in REBVAL* but which were not actually
GC-visible.  Since the series had not been migrated to managed status,
the series in the value cell could conceivably be freed.

The existence of this class of REBVAL meant that the system then had
to worry about making sure it didn't have that kind of value at a lot
of places in the system.  e.g. an evaluation result should always be
a "managed value".  These tests wound up strewn around the system with
the ASSERT_VALUE_MANAGED macro...and wound up adding up to a non
trivial cost of ~5% overhead in the debug build.

This commit eliminates the distinction, and says that all value cells
are managed.  The only remaining case where unmanaged initializations
were needed was in PROBE(), and this adapts it to be able to avoid
using value cells and do direct output based on the series alone.  The
result simplifies the system and improves performance of the debug
build (no impact on release build, however).

Also changes it so that the check that a variable binding's symbol
matches the spelling of the key is not done by default, as it is a
slow check that has never triggered.