Small performance enhancements for debug build
Looking at callgrind showed a few bits of low-hanging fruit to make
the debug build a little bit faster.  One aspect is that the debug
build does not do inlining of functions, so that means an "inline"
function doesn't get inlined...and the more parameters that function
has, the more time will be spent pushing them to the stack.  Some
extremely common debug routines were taking 3 arguments, and called
so often that merely the overhead of calling them (not even what they
were doing) was taking in excess of 1% of the runtime.

An easy way to work around this is just to use "unsafe" macros which
repeat their arguments, which for these cases is fine.

Also, there was some checking against "trash" cells being done on
every evaluator loop.  The system is fairly intolerant of trash cells,
and any access for any other check would catch it...so it's not really
too necessary to specifically check for that.