Eliminate rebTimeHMS(), rebDateYMD(), rebDateTime()
As an API client, ODBC required the making of types for which no C
API to create them existed.  At that early stage of API design, it
made sense to create those routines, e.g.

    REBVAL *time = rebTimeHMS(hour, minute, second);
    REBVAL *date = rebDateYMD(year, month, day);
    REBVAL *datetime = rebDateTime(date, time);

However, as the API vision has become clearer as one of very limited
entry points, the better way is to use Rebol code:

    REBVAL *time = rebRun("make time! [",
        rebI(hour), rebI(minute), rebI(second),
    "]", END);

    REBVAL *date = rebRun("make date! [",
        rebI(day), rebI(month), rebI(year),
    "]", END);

    REBVAL *datetime = rebRun("use [date] [",
        "date:", date,
        "date/time:", time,
    "]", END);

That third one is a little weird, and anyone who wants to suggest a
better way of doing it can chime in here:

https://github.com/rebol/rebol-issues/issues/2313

This commit goes ahead and changes the time and date creation to use
these functions.  There's not currently a way to use integer nano-
seconds in Rebol (though you can MAKE TIME! with a DECIMAL! seconds
component), so for round-trip precision the rebTimeNano() API is
kept as-is.