TO TEXT! of a BLANK! changed to underscore
Some source-level usages of BLANK! are using them because they are the
lexical version of an underscore.  They're being used for their
"underscore-ness", not for their "BLANK!-ness", and the specific
expectation is that they render as an underscore in TO TEXT!, PRINT,
and other string contexts.

Also gets rid of lone # as a way of making a BLANK!.  The idea that
# was NONE! came from R3-Alpha, and didn't exist in Rebol2 (nor was
it picked up by Red).  In Rebol2 it had been an empty ISSUE!, but
that's not currently technically allowed since ISSUE! is now an
ANY-WORD!.

So for now # is just an error (which it is in Red, too).  It should
be reviewed in the future if it has another meaning, but since it
should certainly not be a BLANK! it's good to go ahead and remove
that meaning now.