NAY => NOT-THEN, tweak AND[...]/OR[...] to return NULL
In considering what distinguishes the short-circuiting nature of AND
when it has a BLOCK! on the right from THEN, the main distinction
is that THEN needs to give a BLANK! instead of a NULL if it ever
runs the right hand side.  Otherwise ELSE cannot be "telegraphed"
whether the condition on the left was truthy or not.  An AND does
not want to do this, as any falsey right hand side evaluation should
yield the same result as any falsey left hand side... a NULL.
This effectively makes it more like an ALL which takes exactly two
expressions, vs. like a THEN.

This enforces that invariant.  Also, as THEN is infix IF, there had
been a question of what infix old-style UNLESS would be.  No good name
had been come up with, so NAY was used as a placeholder.  In the time
since, UNLESS has been repurposed and so it really is just IF-NOT for
the native that is the complement to IF.  So NOT-THEN makes more sense
than NAY...though people probably won't bother using it much.