Eliminate RDC_POLL and RDO_AUTO_POLL
It is a goal to basically eliminate the "device" model, and move to
an extension model where PORT!s express basically all of their
extensibility through usermode Rebol interfaces.

This removes the "custom polling" hook from devices.  What used to
happen was that the function behind WAIT would go through and poll all
the devices to see if they had any requests which were asynchronous,
-but- a device could overload this with something that would look
through the request list itself.  The only place this feature was
really used was asynchronous DNS, which was removed (the feature was
only available on Windows, but the function providing it had been
deprecated and would not be supported for IPv6...you were expected
to make your own thread and use standard DNS querying).

It also removes the class of device option which would "poll device,
even if no requests".  There were no instances of this, and if someone
needed to accomplish it, they could just put in a request that never
says it's done...but kept asking to be left in the pending queue.