Have network device add PORT! to connections directly
In R3-Alpha, the desire for the socket layer to not be aware of Rebol
datatypes meant that a listening socket accepting a client connection
went through two steps.  The first step was that the connection was
accepted to build a temporary malloc()'d REBREQ structure, which was
kept in a singly linked list in a variable called `sock`.  (The link
was via the ->next field in the REBREQ, which was generally used for
polling purposes...but semi-"abused" in reuse for the purposes of
this list).  Then, the FIRST operator was used in the client's ACCEPT
event handling, which is where the state would actually become turned
into a PORT!, free()'ing the temporary structure.

Ren-C is abandoning the notion that there is any Rebol layer that
doesn't have at least libRebol available.  So the `sock` field of
REBREQ has been removed, and instead the PORT! object has `connections`
as a BLOCK!.  Now the accept event can add directly to that list,
bypassing the temporary malloc()'d structure.

(Since the `connections` field is only used by the network port, it
really should be more specifically for that.  But this is an interim
step to make sure the connections list is monitored by the GC, until
the port state becomes a Rebol-structured OBJECT!, as opposed to a
C structure...)