Fix GC bug for SERIES_FLAG_INACCESSIBLE arrays
It was happening that if a frame was off the stack, it was not having
its keysource marked.  This is part of a larger set of questions of
how many operations are going to be legal on expired frames, or FREE'd
functions or arrays.  (Should you still be able to tell what file and
line an array had been at?  Can two FRAME!s be compared for identity?
Can you tell what ACTION! an expired frame *had been* for?  etc.)

For now this just goes ahead and lets the expired frame mark its
keysource as live.  However, it should be reconsidered.

Includes another bugfix for events, which were overwriting the format
of the cell that MAKE EVENT! was writing into.